+++ translations/traceview_zh_CN.ts
@@ -586,34 +586,91 @@
 <context>
     <name>traceview::AboutDialog</name>
     <message>
-        <location filename="../lib/core/aboutdialog.cpp" line="+13"/>
+        <location filename="../lib/core/aboutdialog.cpp" line="+24"/>
         <source>About TraceView</source>
         <translation>关于 TraceView</translation>
     </message>
     <message>
+        <location line="+16"/>
+        <source>Version:</source>
+        <translation>版本：</translation>
+    </message>
+    <message>
+        <location line="+1"/>
+        <source>Platform:</source>
+        <translation>平台：</translation>
+    </message>
+    <message>
         <location line="+3"/>
-        <source>&lt;b&gt;TraceView&lt;/b&gt; v%1</source>
-        <translation>&lt;b&gt;TraceView&lt;/b&gt; v%1</translation>
+        <source>Installed from:</source>
+        <translation>安装来源：</translation>
+    </message>
+    <message>
+        <location line="+1"/>
+        <source>Qt:</source>
+        <translation>Qt：</translation>
+    </message>
+    <message>
+        <location line="+1"/>
+        <source>Developer:</source>
+        <translation>开发者：</translation>
     </message>
     <message>
         <location line="+2"/>
+        <source>Contact:</source>
+        <translation>联系方式：</translation>
+    </message>
+    <message>
+        <location line="+5"/>
+        <source>License:</source>
+        <translation>许可证：</translation>
+    </message>
+    <message>
+        <location line="+18"/>
+        <source>Flatpak (alisontristao.github.io/TraceView)</source>
+        <translation>Flatpak (alisontristao.github.io/TraceView)</translation>
+    </message>
+    <message>
+        <location line="+2"/>
+        <source>APK from GitHub Releases</source>
+        <translation>来自 GitHub Releases 的 APK</translation>
+    </message>
+    <message>
+        <location line="+5"/>
+        <source>Windows installer from GitHub Releases</source>
+        <translation>来自 GitHub Releases 的 Windows 安装程序</translation>
+    </message>
+    <message>
+        <location line="+2"/>
+        <location line="+5"/>
+        <location line="+2"/>
+        <source>Local build</source>
+        <translation>本地构建</translation>
+    </message>
+    <message>
+        <location line="-4"/>
+        <source>AppImage from GitHub Releases</source>
+        <translation>来自 GitHub Releases 的 AppImage</translation>
+    </message>
+    <message>
+        <source>&lt;b&gt;TraceView&lt;/b&gt; v%1</source>
+        <translation type="vanished">&lt;b&gt;TraceView&lt;/b&gt; v%1</translation>
+    </message>
+    <message>
         <source>Real-time telemetry dashboard for ESP32/ESP-NOW robots</source>
-        <translation>面向 ESP32/ESP-NOW 机器人的实时遥测仪表盘</translation>
+        <translation type="vanished">面向 ESP32/ESP-NOW 机器人的实时遥测仪表盘</translation>
     </message>
     <message>
-        <location line="+4"/>
         <source>Built with Qt %1 &amp;middot; running with Qt %2</source>
-        <translation>使用 Qt %1 构建 &amp;middot; 使用 Qt %2 运行</translation>
+        <translation type="vanished">使用 Qt %1 构建 &amp;middot; 使用 Qt %2 运行</translation>
     </message>
     <message>
-        <location line="+3"/>
         <source>MIT License &amp;middot; AlisonTristao</source>
-        <translation>MIT 许可证 &amp;middot; AlisonTristao</translation>
+        <translation type="vanished">MIT 许可证 &amp;middot; AlisonTristao</translation>
     </message>
     <message>
-        <location line="+3"/>
         <source>See CHANGELOG.md for release history.</source>
-        <translation>版本历史请参见 CHANGELOG.md。</translation>
+        <translation type="vanished">版本历史请参见 CHANGELOG.md。</translation>
     </message>
 </context>
 <context>
@@ -2038,7 +2095,7 @@ RTT: %3 ms</translation>
 <context>
     <name>traceview::LoginDialog</name>
     <message>
-        <location filename="../lib/core/logindialog.cpp" line="+15"/>
+        <location filename="../lib/core/logindialog.cpp" line="+17"/>
         <source>Enter Developer Mode</source>
         <translation>进入开发者模式</translation>
     </message>
@@ -4074,6 +4131,11 @@ Choose Later to keep working -- the changes will apply the next time you open Tr
         <location line="+2"/>
         <source>&lt;b&gt;Update available&lt;/b&gt;</source>
         <translation>&lt;b&gt;有可用更新&lt;/b&gt;</translation>
+    </message>
+    <message>
+        <location line="+4"/>
+        <source>A new update is available. Do you want to update?</source>
+        <translation>有新的更新可用。是否立即更新？</translation>
     </message>
     <message>
         <location line="+4"/>